the brach finishes the directory refactoring the app dir containts the api and the tests directories the instance contains the config file for the app the root dir contains the run.py,travis,readme and requirements files [finishes #160728715]
