+++ .travis.yml
@@ -8,7 +8,7 @@ install:
   - pip install coveralls
 
 script:
-  - nosetests --with-coverage --cover-package=app
+  - nosetests --with-coverage --cover-package=app/tests/v1
 
 after_success:
   - coveralls
